inputLayer/requirements.txt
# Trademark PDF Adaptive Parser — Requirements
# Install with: pip install -r requirements.txt

# PDF extraction
pdfplumber==0.11.4

# HTTP client for Groq API
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# UI
streamlit>=1.35.0

# ── Optional: if you later add OCR for scanned PDFs ──────
# pytesseract>=0.3.10
# pdf2image>=1.17.0
